% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNetLogoData.R
\name{barcode.timecourse}
\alias{barcode.timecourse}
\title{Plot abundance of barcode lineages over time.}
\usage{
barcode.timecourse(NLdata, runnum)
}
\arguments{
\item{NLdata}{The full dataframe containing data from all BehaviorSpace
experiments at all timepoints, imported from NetLogo.}

\item{runnum}{The run number of the experiment of interest.}
}
\value{
A ggplot object using [ggTimeSeries::stat_steamgraph()].
}
\description{
Generates a streamgraph plot displaying the abundance of each unique barcode
lineage as a proportion of the population, over time, in a Pheno-Evo
experiment. At the initiation of the experiment, each founding cell is given
a unique barcode. A barcode lineage consists of a founding cell and all its
descendants. Over time, especially with high population turnover, some
barcodes go extinct. This plot shows barcodes as different colors, with their
proportional abundance as their width along the y-axis, and time on the
x-axis.
}
\details{
There is not a unique color for each barcode. This plot uses a 12-color
palette that is repeated several times to cover all the barcodes. However,
each lineage "stream" is bordered by a black line, so if two lineages that
happen to have the same color end up adjacent, they can be distinguished by
the border between them.
}
\examples{
data(PhenoEvoData_4)
barcode.timecourse(PhenoEvoData_run4, 4)
}
