% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNetLogoData.R
\name{cell.abundance.timecourse}
\alias{cell.abundance.timecourse}
\title{Plot total cell population and toxin concentration over time}
\usage{
cell.abundance.timecourse(NLdata, runnum)
}
\arguments{
\item{NLdata}{The full dataframe containing data from all BehaviorSpace
experiments at all timepoints, imported from NetLogo.}

\item{runnum}{The run number of the experiment of interest.}
}
\value{
A ggplot object using geom_line. This can be modified in the typical
  ggplot way by adding layers, scales, etc.
}
\description{
Produces a simple line plot showing the total population of cells in red, and
the average concentration of toxin per patch in black, over time.
}
\details{
As this function is usually used for the purposes of displaying general
population behavior, results may be abridged: if the input dataframe has more
than 1,000 timepoints, only the final 500 timepoints are plotted. An attempt
is made to scale the y-axis of the cell abundance plot so that it can be
displayed on the same plot as the toxin plot.
}
\examples{
data(PhenoEvoData_4)
cell.abundance.timecourse(PhenoEvoData_run4, 4)
}
