% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNetLogoData.R
\name{degrade.rate.heatmap}
\alias{degrade.rate.heatmap}
\title{Heatmap showing average phenotypic degrade rate in many Pheno-Evo populations}
\usage{
degrade.rate.heatmap(ends.df, xvar, yvar)
}
\arguments{
\item{ends.df}{Dataframe of endpoint data, generated by extract.endpoint().
Must contain a degrade.rate.mean column (use [summarize.endpoint()]}

\item{xvar}{Experimental parameter for plotting on x-axis. Must be a column
name of ends.df, with no quotation marks.}

\item{yvar}{Experimental parameter for plotting on y-axis. Must be a column
name of ends.df, with no quotation marks.}
}
\value{
A ggplot object using geom_tile. This can be modified in the typical
  ggplot way by adding layers, scales, etc.
}
\description{
This plot shows three dimensions of data: two experimental parameters (for
instance, toxin concentration and diffusion rate) as axes, and the population
average degrade rate as color gradient from light to dark blue. It is a
snapshot of several experiments from a parameter sweep, at a single
timepoint.
}
\details{
As with other endpoint-summary plots, this function can be used with data
from any timepoint, as long as the ends.df dataframe contains data from only
one timepoint per run number. It is a snapshot of several experiments from a
parameter sweep, at a single timepoint. For a function to create a heatmap
with any variable you wish (instead of degrade rate), see
[phenoevo.heatmap()].
}
\examples{
data(PE.ends)
degrade.rate.heatmap(PE.ends, toxin.conc, env.noise)
}
