% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndpointSummaryStats.R
\name{degraderate.variogram}
\alias{degraderate.variogram}
\title{Calculate the variogram for degrade.rate for a given experiment}
\usage{
degraderate.variogram(ends.df, run.num)
}
\arguments{
\item{ends.df}{Dataframe of endpoint data, generated by extract.endpoint().
Must contain an xydr column.}

\item{run.num}{Run number to be analyzed.}
}
\value{
An object of class gstatVariogram. See [gstat::variogram()]
  for documentation
}
\description{
This function adapts Pheno-Evo data and then uses
[gstat::variogram()] from the gstat package to calculate spatial
correlation in the phenotypic degrade rate values of the cells. It returns a
dataframe. For a variogram plot, call plot() on the dataframe - see examples.
}
\details{
Dataframes from timepoints other than the endpoint are also fine, as long as
there is only 1 row per run number.
}
\examples{
data(PE.ends)
degraderate.variogram(PE.ends, 4)
plot(degraderate.variogram(PE.ends, 4))
}
