% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndpointSummaryStats.R
\name{extract.endpoint}
\alias{extract.endpoint}
\title{Extract the final timepoints from a set of NetLogo experiment runs}
\usage{
extract.endpoint(NLdata, run.nums)
}
\arguments{
\item{NLdata}{The full dataframe containing data from all BehaviorSpace
experiments, imported from NetLogo.}

\item{run.nums}{A vector of the run numbers given to the experiments by
BehaviorSpace.}
}
\value{
A new dataframe with all the data from just the final timepoints of
  all experiments.
}
\description{
Creates a new dataframe for easy analysis comparing multiple Pheno-Evo model
runs from a parameter sweep using NetLogo BehaviorSpace, by storing only the
end results of each model run. If the experiments in your BehaviorSpace run
lasted for differing lengths of time (e.g., if some populations went extinct
rapidly and those experiments ended prematurely), this function will figure
out what the last timepoint was that had a nonzero population, and extract
it.
}
\details{
It is fairly important to make sure that the column names are standardized as
described in the Pheno-Evo tutorial.
}
\examples{
data(PhenoEvoData_small)
PE.ends_small<-extract.endpoint(PhenoEvoData_small, c(2, 4, 6))
}
