% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseNetlogoData.R
\name{extractXYdr}
\alias{extractXYdr}
\title{Interpret the "xydr" variable from Pheno-Evo model runs}
\usage{
extractXYdr(charstring)
}
\arguments{
\item{charstring}{The full character string from one entry in the xydir column
of a NetLogo data table, including brackets and quotation marks.}
}
\value{
A dataframe with three columns: xcor, ycor, and degrade.rate.
Each row represents the data from one cell.
}
\description{
In Pheno-Evo, xydr is a property of a cell that is a list of the
x-coordinate, the y-coordinate, and the degrade rate. These are reported
together in order to enable spatial analysis of degrade rate values, which
would not otherwise be possible because coordinate data would be separated from degrade rate data.
[extractXYdr()] takes all the xydr values reported in one entry
(one timepoint of one experiment) of a Pheno-Evo data table and separates
them out into a new dataframe with appropriately labeled columns.
}
\details{
In general, the normal Pheno-Evo user should not have to use this function.
It is called by several other functions used for analyzing Pheno-Evo
output data, such as [degraderate.variogram()].
}
\examples{
data(PE.ends)
PE.ends$xydr[4]
extractXYdr(PE.ends$xydr[4])
}
