% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNetLogoData.R
\name{generation.heatmap}
\alias{generation.heatmap}
\title{Heatmap showing average number of generations in many Pheno-Evo populations}
\usage{
generation.heatmap(ends.df, xvar, yvar)
}
\arguments{
\item{ends.df}{Dataframe of endpoint data, generated by extract.endpoint().
Must contain a generation.mean column (use [summarize.endpoint()])}

\item{xvar}{Experimental parameter for plotting on x-axis. Must be a column
name of ends.df, with no quotation marks.}

\item{yvar}{Experimental parameter for plotting on y-axis. Must be a column
name of ends.df, with no quotation marks.}
}
\value{
A ggplot object using geom_tile. This can be modified in the typical
  ggplot way by adding layers, scales, etc.
}
\description{
Each cell in the Pheno-Evo model carries a number indicating the number of
generations elapsed between the founding ancestor cells and its birth. At a
given timepoint, the average of the generation numbers of all cells in the
population can serve as a rough indicator of the population's overall growth
rate. [generation.heatmap()] shows three dimensions of data: two
experimental parameters (for instance, toxin concentration and diffusion
rate) as axes, and the population average generation number as color gradient
from pale yellow to dark green. It is a snapshot of several experiments from
a parameter sweep at a single timepoint.
}
\details{
As with other endpoint-summary plots, this function can be used with data
from any timepoint, as long as the ends.df dataframe contains data from only
one timepoint per run number. For a function to create a heatmap with any
variable you wish (instead of generation number), see
[phenoevo.heatmap()].
}
\examples{
data(PE.ends)
generation.heatmap(PE.ends, toxin.conc, env.noise)
}
