% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseNetlogoData.R
\name{parsefactors.norm}
\alias{parsefactors.norm}
\title{Factors from the character values from a single cell of a NetLogo output
table}
\usage{
parsefactors.norm(charstring)
}
\arguments{
\item{charstring}{The full character string from one entry of a NetLogo data
table, including brackets and quotation marks.}
}
\value{
A dataframe with three columns: s3 = factor levels; Freq = number of
  individuals at each factor level; Prop = proportion of the total community.
}
\description{
In Pheno-Evo experiments, it is common to store all the values from all
agents for a particular variable at each given timepoint. In the resulting
Netlogo data table, those values all go into a single entry of the table at
each timepoint. [parsenfactors.norm()] can be used to split up
those values and export them into a new dataframe as factors, along with
information about their frequency in the population. As an example: the list
of individual barcodes carried by cells in a Pheno-Evo population.
}
\details{
In general, the normal Pheno-Evo user should not have to use this function.
It is called by several other functions used for analyzing Pheno-Evo output
data, such as [barcode.timecourse()]. For a function that
calculates density and returns a dataframe, see [parsenumeric()].
For a function that returns a vector of numeric values, see
[extractnumeric()].
}
\examples{
data(PE.ends)
PE.ends$barcode[4]
parsefactors.norm(PE.ends$barcode[4])
}
