% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseNetlogoData.R
\name{parsenumeric}
\alias{parsenumeric}
\title{Histogram from the numeric values from a single cell of a NetLogo output
table}
\usage{
parsenumeric(charstring)
}
\arguments{
\item{charstring}{The full character string from one entry of a NetLogo data
table, including brackets and quotation marks.}
}
\value{
A dataframe with a single column named s4, giving population density
  estimates across the range of the trait value from 0 and 1.00 in bins of
  width 0.01.
}
\description{
In Pheno-Evo experiments, it is common to store all the values from all
agents for a particular variable at each given timepoint. In the resulting
Netlogo data table, those values all go into a single entry of the table at
each timepoint. [parsenumeric()] can be used to split up those
values and export them into a new dataframe - specifically, it returns a
histogram of the values.
}
\details{
In general, the normal Pheno-Evo user should not have to use this function.
It is called by several other functions used for analyzing Pheno-Evo output
data, such as [calc.fun.div()] and
[phenotype.histogram()]. For a function that returns the vector of
numbers and does not calculate density, see [extractnumeric()]. For
a function that returns the values as factors rather than numeric, see
[parsefactors.norm()].
}
\examples{
data(PE.ends)
PE.ends$degrade.rate[4]
parsenumeric(PE.ends$degrade.rate[4])
}
