% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNetLogoData.R
\name{response.error.timecourse}
\alias{response.error.timecourse}
\title{Plot population distribution of response.error over time}
\usage{
response.error.timecourse(NLdata, runnum)
}
\arguments{
\item{NLdata}{The full dataframe containing data from all BehaviorSpace
experiments at all timepoints, imported from NetLogo.}

\item{runnum}{The run number of the experiment of interest.}
}
\value{
A ggplot object using geom_tile. This can be modified in the typical
  ggplot way by adding layers, scales, etc.
}
\description{
Generates a heatmap displaying how the genotypic response error changes over
the course of a Pheno-Evo experiment. Time is displayed on the x-axis, the
range of possible response error values on the y-axis, and relative abundance
given by color. To reduce plot size and complexity, if the input dataframe contains
more than 1,000 timepoints, only every 10th timepoint is plotted.
}
\details{
See also [switch.rate.timecourse()] and [degrade.rate.timecourse()].
}
\examples{
data(PhenoEvoData_4)
response.error.timecourse(PhenoEvoData_run4, 4)
}
