% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNetLogoData.R
\name{survival.heatmap}
\alias{survival.heatmap}
\title{Heatmap of population survival duration in NetLogo experiments}
\usage{
survival.heatmap(ends.df, xvar, yvar, nums = FALSE)
}
\arguments{
\item{ends.df}{Dataframe of endpoint data, generated by extract.endpoint().}

\item{xvar}{Experimental parameter for plotting on x-axis. Must be a column
name of ends.df, with no quotation marks.}

\item{yvar}{Experimental parameter for plotting on y-axis. Must be a column
name of ends.df, with no quotation marks.}

\item{nums}{Should each box in the heatmap be labeled with its run number?
T/F. Default: F}
}
\value{
A ggplot object using geom_tile. This can be modified in the typical
  ggplot way by adding layers, scales, etc.
}
\description{
Generates a plot with three dimensions of data: two experimental parameters
(for instance, toxin concentration and diffusion rate) as axes, and the
number of timepoints that each population survived before extinction given as
a color gradient from pale to dark red. It is a snapshot of several
experiments from a parameter sweep, at a single timepoint. For ease of
interpretation, the boxes in the heatmap can be labeled by run number.
}
\details{
As with other endpoint-summary plots, this function can be used with data
from any timepoint, as long as the ends.df dataframe contains data from only
one timepoint per run number. For a function to create a heatmap with any
variable you wish (instead of population survival), see
[phenoevo.heatmap()].
}
\examples{
data(PE.ends)
survival.heatmap(PE.ends, toxin.conc, env.noise, nums=TRUE)
}
