% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNetLogoData.R
\name{switch.rate.timecourse}
\alias{switch.rate.timecourse}
\title{Plot population distribution of switch.rate over time}
\usage{
switch.rate.timecourse(NLdata, runnum)
}
\arguments{
\item{NLdata}{The full dataframe containing data from all BehaviorSpace
experiments at all timepoints, imported from NetLogo.}

\item{runnum}{The run number of the experiment of interest.}
}
\value{
A ggplot object using geom_tile. This can be modified in the typical
  ggplot way by adding layers, scales, etc.
}
\description{
Generates a heatmap displaying how the genotypic switch rate changes over the
course of a Pheno-Evo experiment. Time is displayed on the x-axis, the range
of possible switch rate values on the y-axis, and relative abundance given by
color. To reduce plot size and complexity, if the input dataframe contains
more than 1,000 timepoints, only every 10th timepoint is plotted.
}
\details{
#' See also [degrade.rate.timecourse()] and [response.error.timecourse()].
}
\examples{
data(PhenoEvoData_4)
switch.rate.timecourse(PhenoEvoData_run4, 4)
}
